backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-filter==25.2
psycopg2-binary==2.9.11
python-dotenv==1.0.1
gunicorn==21.2.0
